docs: Additional features for MUI chrome in nextjs app.
