+++ examples/nextjs-with-typescript/src/app/material-ui-player-chrome/components/FullscreenButton.tsx
@@ -1,4 +1,4 @@
-import { IconButton } from '@mui/material';
+import { IconButton, Tooltip } from '@mui/material';
 import FullscreenIcon from '@mui/icons-material/Fullscreen';
 import FullscreenExitIcon from '@mui/icons-material/FullscreenExit';
 import {
@@ -13,19 +13,22 @@ const FullscreenButton = () => {
     (state) => state.mediaIsFullscreen
   );
   const IconComponent = mediaIsFullscreen ? FullscreenExitIcon : FullscreenIcon;
+  const label = mediaIsFullscreen ? 'Exit full screen' : 'Full screen';
   return (
-    <IconButton
-      aria-label={mediaIsFullscreen ? 'exit fullscreen' : 'enter fullscreen'}
-      color="primary"
-      onClick={() => {
-        const type = mediaIsFullscreen
-          ? MediaActionTypes.MEDIA_EXIT_FULLSCREEN_REQUEST
-          : MediaActionTypes.MEDIA_ENTER_FULLSCREEN_REQUEST;
-        dispatch({ type });
-      }}
-    >
-      <IconComponent />
-    </IconButton>
+    <Tooltip title={label} placement="top">
+      <IconButton
+        aria-label={label}
+        color="primary"
+        onClick={() => {
+          const type = mediaIsFullscreen
+            ? MediaActionTypes.MEDIA_EXIT_FULLSCREEN_REQUEST
+            : MediaActionTypes.MEDIA_ENTER_FULLSCREEN_REQUEST;
+          dispatch({ type });
+        }}
+      >
+        <IconComponent />
+      </IconButton>
+    </Tooltip>
   );
 };
