+++ examples/nextjs-with-typescript/src/app/material-ui-player-chrome/components/PipButton.tsx
@@ -1,4 +1,4 @@
-import { IconButton } from '@mui/material';
+import { IconButton, Tooltip } from '@mui/material';
 import PictureInPictureIcon from '@mui/icons-material/PictureInPicture';
 import {
   useMediaDispatch,
@@ -9,21 +9,22 @@ import {
 const PipButton = () => {
   const dispatch = useMediaDispatch();
   const mediaIsPip = useMediaSelector((state) => state.mediaIsPip);
+  const label = mediaIsPip ? 'Exit picture in picture' : 'Picture in picture';
   return (
-    <IconButton
-      aria-label={
-        mediaIsPip ? 'exit picture in picture' : 'enter picture in picture'
-      }
-      color="primary"
-      onClick={() => {
-        const type = mediaIsPip
-          ? MediaActionTypes.MEDIA_EXIT_PIP_REQUEST
-          : MediaActionTypes.MEDIA_ENTER_PIP_REQUEST;
-        dispatch({ type });
-      }}
-    >
-      <PictureInPictureIcon />
-    </IconButton>
+    <Tooltip title={label} placement="top">
+      <IconButton
+        aria-label={label}
+        color="primary"
+        onClick={() => {
+          const type = mediaIsPip
+            ? MediaActionTypes.MEDIA_EXIT_PIP_REQUEST
+            : MediaActionTypes.MEDIA_ENTER_PIP_REQUEST;
+          dispatch({ type });
+        }}
+      >
+        <PictureInPictureIcon />
+      </IconButton>
+    </Tooltip>
   );
 };
