+++ examples/nextjs-with-typescript/src/app/material-ui-player-chrome/components/PlayButton.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { IconButton } from '@mui/material';
+import { IconButton, Tooltip } from '@mui/material';
 import PlayArrowIcon from '@mui/icons-material/PlayArrow';
 import PauseIcon from '@mui/icons-material/Pause';
 import {
@@ -10,21 +10,26 @@ import {
 
 const PlayButton = () => {
   const dispatch = useMediaDispatch();
-  const mediaPaused = useMediaSelector((state) => state.mediaPaused);
-  const IconComponent = typeof mediaPaused !== 'boolean' || mediaPaused ? PlayArrowIcon : PauseIcon;
+  const mediaPaused = useMediaSelector(
+    (state) => typeof state.mediaPaused !== 'boolean' || state.mediaPaused
+  );
+  const IconComponent = mediaPaused ? PlayArrowIcon : PauseIcon;
+  const label = mediaPaused ? 'Play' : 'Pause';
   return (
-    <IconButton
-      aria-label={mediaPaused ? 'play' : 'pause'}
-      color="primary"
-      onClick={() => {
-        const type = mediaPaused
-          ? MediaActionTypes.MEDIA_PLAY_REQUEST
-          : MediaActionTypes.MEDIA_PAUSE_REQUEST;
-        dispatch({ type });
-      }}
-    >
-      <IconComponent />
-    </IconButton>
+    <Tooltip title={label} placement="top">
+      <IconButton
+        aria-label={label}
+        color="primary"
+        onClick={() => {
+          const type = mediaPaused
+            ? MediaActionTypes.MEDIA_PLAY_REQUEST
+            : MediaActionTypes.MEDIA_PAUSE_REQUEST;
+          dispatch({ type });
+        }}
+      >
+        <IconComponent />
+      </IconButton>
+    </Tooltip>
   );
 };
