+++ examples/nextjs-with-typescript/src/app/material-ui-player-chrome/components/SeekForwardButton.tsx
@@ -1,4 +1,4 @@
-import { IconButton } from '@mui/material';
+import { IconButton, Tooltip } from '@mui/material';
 import Forward30Icon from '@mui/icons-material/Forward30';
 import {
   useMediaDispatch,
@@ -6,21 +6,25 @@ import {
   MediaActionTypes,
 } from 'media-chrome/react/media-store';
 
+const label = 'Seek backward';
+
 const SeekForwardButton = () => {
   const dispatch = useMediaDispatch();
   const mediaCurrentTime = useMediaSelector((state) => state.mediaCurrentTime);
   return (
-    <IconButton
-      aria-label="seek forward"
-      color="primary"
-      onClick={() => {
-        const type = MediaActionTypes.MEDIA_SEEK_REQUEST;
-        const detail = (mediaCurrentTime ?? 0) + 30;
-        dispatch({ type, detail });
-      }}
-    >
-      <Forward30Icon />
-    </IconButton>
+    <Tooltip title={label} placement="top">
+      <IconButton
+        aria-label={label}
+        color="primary"
+        onClick={() => {
+          const type = MediaActionTypes.MEDIA_SEEK_REQUEST;
+          const detail = (mediaCurrentTime ?? 0) + 30;
+          dispatch({ type, detail });
+        }}
+      >
+        <Forward30Icon />
+      </IconButton>
+    </Tooltip>
   );
 };
