+++ examples/nextjs-with-typescript/src/app/material-ui-player-chrome/components/Seekbar.tsx
@@ -24,7 +24,7 @@ const Seekbar = () => {
       sx={{
         mx: 2,
       }}
-    ></Slider>
+    />
   );
 };
