--- examples/nextjs-with-typescript/src/app/material-ui-player-chrome/components/SoundControl.tsx
@@ -1,56 +1,0 @@
-import * as React from 'react';
-import Popover from '@mui/material/Popover';
-import Typography from '@mui/material/Typography';
-import { Box } from '@mui/material';
-import MuteButton from './MuteButton';
-import VolumeSlider from './VolumeSlider';
-
-const SoundControl = () => {
-  const [anchorEl, setAnchorEl] = React.useState<HTMLElement | null>(null);
-
-  const handlePopoverOpen = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handlePopoverClose = () => {
-    setAnchorEl(null);
-  };
-
-  const open = Boolean(anchorEl);
-
-  return (
-    <>
-      <Box
-        aria-owns={open ? 'mouse-over-popover' : undefined}
-        aria-haspopup="true"
-        onMouseEnter={handlePopoverOpen}
-        onMouseLeave={handlePopoverClose}
-      >
-        <MuteButton />
-      </Box>
-      <Popover
-        id="mouse-over-popover"
-        sx={{
-          pointerEvents: 'none',
-        }}
-        open={open}
-        anchorEl={anchorEl}
-        anchorOrigin={{
-          vertical: 'top',
-          horizontal: 'left',
-        }}
-        transformOrigin={{
-          vertical: 'bottom',
-          horizontal: 'left',
-        }}
-        onClose={handlePopoverClose}
-      >
-        <Box>
-          <VolumeSlider orientation="vertical" />
-        </Box>
-      </Popover>
-    </>
-  );
-};
-
-export default SoundControl;
