+++ examples/nextjs-with-typescript/src/app/material-ui-player-chrome/components/VolumeSlider.tsx
@@ -1,9 +1,11 @@
-import { Slider } from '@mui/material';
+import { Slider, Tooltip } from '@mui/material';
 import {
   useMediaDispatch,
   useMediaSelector,
   MediaActionTypes,
 } from 'media-chrome/react/media-store';
+
+const label = 'Volume';
 
 const VolumeSlider = ({
   orientation = 'horizontal',
@@ -15,23 +17,25 @@ const VolumeSlider = ({
   const mediaMuted = useMediaSelector((state) => state.mediaMuted);
   const value = mediaMuted ? 0 : (mediaVolume ?? 0.5) * 100;
   return (
-    <Slider
-      aria-label="volume"
-      min={0}
-      max={100}
-      value={value}
-      onChange={(_event, value) => {
-        const type = MediaActionTypes.MEDIA_VOLUME_REQUEST;
-        const detail = (value as number) / 100;
-        dispatch({ type, detail });
-      }}
-      orientation={orientation}
-      size="small"
-      sx={{
-        mx: 2,
-        width: 100,
-      }}
-    ></Slider>
+    <Tooltip title={label} placement="top">
+      <Slider
+        aria-label={label}
+        min={0}
+        max={100}
+        value={value}
+        onChange={(_event, value) => {
+          const type = MediaActionTypes.MEDIA_VOLUME_REQUEST;
+          const detail = (value as number) / 100;
+          dispatch({ type, detail });
+        }}
+        orientation={orientation}
+        size="small"
+        sx={{
+          mx: 2,
+          width: 100,
+        }}
+      />
+    </Tooltip>
   );
 };
