+++ examples/nextjs-with-typescript/src/app/material-ui-player-chrome/page.tsx
@@ -1,9 +1,9 @@
-import { ReactPlayer } from "@/app/material-ui-player-chrome/ReactPlayer";
+import Player from "@/app/material-ui-player-chrome/components/Player";
 
 export default function Page() {
   return (
     <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      <ReactPlayer src="https://stream.mux.com/DS00Spx1CV902MCtPj5WknGlR102V5HFkDe/high.mp4" />
+      <Player src="https://stream.mux.com/DS00Spx1CV902MCtPj5WknGlR102V5HFkDe/high.mp4" />
     </main>
   );
 }
